SNS 페이지의 Fab 버튼(SNS 포스트를 작성하는 페이지로 이동하는)이 사라졌던 현상 해결
이 과정에서 getSafeStringFromQuery, getSafeNumberFromQuery 함수를 외부 파일(유틸 함수)로 분리
